requirements.txt
tweepy
atproto
python-dotenv